Fixed KeePassXC Config Copy Process

+++ modules/home/users/puppy.nix
@@ -26,8 +26,9 @@
       '';
       createKeePassXcConfigSymLink = lib.hm.dag.entryAfter ["writeBoundary"] ''
         $DRY_RUN_CMD mkdir -p "$HOME/.config/keepassxc"
-        $DRY_RUN_CMD rm "$HOME/.config/keepassxc/keepassxc.ini"
-        $DRY_RUN_CMD cp "/etc/nixos/files/config/keepassxc/keepassxc.ini" "$HOME/.config/keepassxc/keepassxc.ini"
+        if [ ! -f "$HOME/.config/keepassxc/keepassxc.ini" ]; then
+          $DRY_RUN_CMD cp "/etc/nixos/files/config/keepassxc/keepassxc.ini" "$HOME/.config/keepassxc/keepassxc.ini"
+        fi
       '';
       createPicomUserConfFile = lib.hm.dag.entryAfter ["writeBoundary"] ''
         $DRY_RUN_CMD mkdir -p "$HOME/.config/picom"
